[busybox] fix exFat automount

+++ meta-oe/recipes-core/busybox/busybox-1.23.2/mdev-mount.sh
@@ -127,7 +127,9 @@ case "$ACTION" in
 				mkdir -p $MOUNTPOINT
 			fi
 			if ! mount -t auto /dev/$MDEV "${MOUNTPOINT}" ; then
-				rmdir "${MOUNTPOINT}"
+				if ! mount.exfat /dev/$MDEV "${MOUNTPOINT}" ; then
+					rmdir "${MOUNTPOINT}"
+				fi
 			fi
 		fi
 		;;
